Refactored SSH server out of lib
